script04

+++ unidade05/script04/combates.sh
@@ -5,7 +5,7 @@ validacion(){
     then 
         echo "O script $0 precisa que se indique o nome do ficheiro de combates";
         exit;
-    elif  [-r $1  ]
+    elif  [ -r $1 -a  -f $1 ] 
     then
         echo "ERRO!! O ficheiro de >$1< non existe ou non ten permisos de lectura";
         exit;
@@ -25,21 +25,16 @@ validacion(){
 principal() {
    while IFS= read -r linea
    do
-      
-
-      loitador01=$(echo $linea | tr -d '"'|cut -d ';' -f1) ;
-      vida01=$(echo $linea| tr -d '"'| cut -d ';' -f2)  ; 
-
-      loitador02=$(echo $linea | tr -d '"' | cut -d ';' -f3);  
-      vida02=$(echo $linea | tr -d '"' | cut -d ';' -f4) ; 
+      loitador01="$(echo $linea | cut -d ';' -f1)";
+      vida01=$(echo $linea      | cut -d ';' -f2)  ; 
+      loitador02=$(echo $linea  | cut -d ';' -f3);  
+      vida02=$(echo $linea      | cut -d ';' -f4) ; 
 
       echo -e "\nProcesando $linea"; 
-      preparacion $loitador01 $loitador02
-      combate $loitador01 $vida01 $loitador02 $vida02
 
+      preparacion  "${loitador01}" "${loitador02}" 
+      combate "${loitador01}" "${vida01}"  "${loitador02}"  "${vida02}"
 
-     # echo -e  "$loitador01  $vida01\n"
-     # echo -e  "$loitador02 $vida02 \n"
 
     done < $1
 
@@ -48,8 +43,11 @@ principal() {
 
 
 preparacion() {
+   RED='\033[0;31m';
+   NC='\033[0m' ;
+   Purple='\033[0;35m'; 
    trap "echo -e ' \t Saltando a preparación' && break" INT TSTP
-   echo -e "No seguinte combate $1 se enfrontará a $2!!! Quen gañará?"; 
+   echo -e "No seguinte combate ${RED} ${1} ${NC} enfrontarase a ${Purple} ${2} ${NC} !!! Quen gañará?"; 
 
    for ((i=3; i>=1; i--))
    do
@@ -60,11 +58,46 @@ preparacion() {
 }
  
 combate() { 
-   echo "Que comece o combate!!!";
-   echo "$1 pelexa contra $3 $2 $4";
+   trap '' INT TSTP
 
-   #ataque
-   #contador
+   #Params: $1=$loitador01 $2=$vida01 $3=$loitador02 $4=$vida02
+   loitador01=$1;
+   vida01=$2;
+   loitador02=$3;
+   vida02=$4;
+
+   VERDE='\033[42m';
+   NC='\033[0m' ;
+
+   echo -e "${VERDE} Que comece o combate!!! ${NC} ";
+   echo "$loitador01 pelexa contra $loitador02 ";
+
+   while [ $vida01 -gt 0 -a $vida02 -gt 0 ]
+   do
+      ataque01=$(shuf -i 1-20 -n 1);
+      ataque02=$(shuf -i 1-20 -n 1);
+      vida02="$(($vida02-$ataque01))"
+      vida01="$(($vida01-$ataque02))"
+
+      echo -e "$loitador01 ataca a $loitador02 e fai $ataque01 puntos de dano."
+      if [ $vida02 -gt 0 ]
+         then   echo -e "$loitador02 recibe o golpe. Quédanlle $vida02 puntos de vida."
+         else   echo -e "$loitador02  perdeu o combate.";
+                break;
+      fi
+      sleep 1;
+
+      echo -e "$loitador02 ataca a $loitador01 e fai $ataque02 puntos de dano."
+      if [ $vida01 -gt 0 ]
+         then   echo -e "$loitador01 recibe o golpe. Quédanlle $vida01 puntos de vida."
+         else   echo -e "$loitador01  perdeu o combate.";
+                break;
+      fi
+      sleep 1;
+      
+   done
+
+   echo -e "Combate finalizado\n"
 
 }
 
@@ -76,10 +109,3 @@ combate() {
 validacion $@
 principal $@
 
-
-# trap '' INT TSTP
-# for ((i=0; i<=1; i++))
-# do
-#    echo "pensando"
-#    sleep 1
-# done
